Remove redundant and outdated library metadata file
library.json is PlatformIO's native library manifest file but library.properties is also supported:
https://docs.platformio.org/en/latest/librarymanager/creating.html

The library.json file has not been maintained. The version is outdated and it also has an incorrect `platforms` value. So the evidence makes it obvious that having two metadata files in this project has a harmful effect. Since the Arduino development software requires library.properties, the choice of keeping library.properties and removing library.json is clear.