Update NTPClient.cpp
Until now the fraction of the second was not taken into account. Therefore the resulting time offset is up to 1000 ms. Using the provided Information increases the accuracy of the synchronization to 20-30 milliseconds. And it requires just two lines of code.

https://imgur.com/a/aVjvHON
In this plot both methods are compared. Plotted is the time difference in respect to a PPS signal provided by a GPS receiver.  For comparison the time was synchronized over 1000 times and the resulting time difference filled into the respective histogram.